feat: enhance binary array encoding, make it the default (#2521)
This adds support for the following things to the binary encoding:
 * Nulls
 * Large offsets
 * Different types (e.g. String vs Binary)

In addition, I have changed the row limit on pages from `u32` to `u64`.
Partly, this is because string arrays larger that 2GB are not unheard of
and it may be nice in some cases to store them in one page. Also pages
are infrequent enough that an extra 4 bytes is trivial (actually,
protobuf already stores u32 and u64 the same). Another reason is that,
with some encodings (e.g. constant / compressed bitmap / etc.) it's
possible to have 2^64 rows without occupying a large amount of space.

Note: this is technically a `.proto` change but in protobuf `u32 -> u64`
is backwards compatible so it is not a huge issue.

+++ rust/lance-encoding/src/encodings/physical/basic.rs
@@ -122,7 +122,7 @@ impl BasicPageScheduler {
 impl PageScheduler for BasicPageScheduler {
     fn schedule_ranges(
         &self,
-        ranges: &[std::ops::Range<u32>],
+        ranges: &[std::ops::Range<u64>],
         scheduler: &Arc<dyn EncodingsIo>,
         top_level_row: u64,
     ) -> BoxFuture<'static, Result<Box<dyn PrimitivePageDecoder>>> {
@@ -171,8 +171,8 @@ struct BasicPageDecoder {
 impl PrimitivePageDecoder for BasicPageDecoder {
     fn decode(
         &self,
-        rows_to_skip: u32,
-        num_rows: u32,
+        rows_to_skip: u64,
+        num_rows: u64,
         all_null: &mut bool,
     ) -> Result<Vec<BytesMut>> {
         let dest_buffers = match &self.mode {
